src/requirements.txt
flask
flask-cors
instagrapi
moviepy
imageio[ffmpeg]
numpy
python-dotenv
pyjwt